chore: roll to Playwright v1.35.1 (#354)
* chore: roll to Playwright v1.35.1

* fix: rpc protocol update

* feat: Locators support chain calls

BREAKING CHANGE: Make Locator easier to chain calls, e.g:
`_, err := page.Locator("body").Locator("div").And("p").TextContent()`

+++ assertions.go
@@ -42,7 +42,7 @@ type expectedTextValue struct {
 type frameExpectOptions struct {
 	ExpressionArg  interface{}         `json:"expressionArg"`
 	ExpectedText   []expectedTextValue `json:"expectedText"`
-	ExpectedNumber *int                `json:"expectedNumber"`
+	ExpectedNumber *float64            `json:"expectedNumber"`
 	ExpectedValue  interface{}         `json:"expectedValue"`
 	UseInnerText   *bool               `json:"useInnerText"`
 	IsNot          bool                `json:"isNot"`
